Initial commit
To pass the trial this repository must include a single dockerfile
that upon builds an image based on Ubuntu. The container of
resulting image must access a .cpp from a remote repository,
compile it, and then run the resulting executable. The direct usage
of git is required, though it's not specified whether it should
occur upon building an image or running a container. It is also
unspecified how the container gets the ability to compile nor when
does the compiling occur.

I suppose that installing the required apps must occur when the image
builds, so that it only happens once and does not require the usage of
docker-compose of corresponding images, which would unnecessarily
complicate the whole process, and possibly increase size of resulting
image.

On the other hand, pulling repository inside the container could occur
only once by using volumes. This would add flexibility to using the
container, since pulling and compiling could be omitted on subsequent
runs, or even the whole process of pulling-compiling-running could
never happen by specifying different command upon running image.

For now the image result of "build ." consists of:
* Ubuntu v22.04
* actual version of git installed
* actual version of g++ compiler installed
* actual version of cmake installed

Important notes:
All installations are made by using unupgraded version of apt.
Apt is not connected to any streamline repositories, so installed
versions of required apps are probably not up-to-date.