Reimagined solution, improved QOL on Dockerfile
List of changes:
1. Added LINKS.MD as to collect reference
2. Added URL.txt as to store target repo URL
3. Renamed Dockerfile to deprecatedDkerfile
4. Created a refactored version of Dockerfile with QOL improvements

Turns out, that target repo is a Sample collection (as in: "The
collection of code snippets to demonstrate work of the actual API")
Turns out, that the "actual API" is massive. As in: "The actual API
is onnxruntime - Open Neural Network Exchange runtime". After a little
digging, I found out that this API is used as a sort of interface to a
whole lot of other API's, both for training and inference. Specifically,
ONNX runtime is used for inference. Anyway, the specific snippet, that
has to be compiled depends on ONNX runtime. ONNX runtime, in turn,
depends on cuDNN. Which depends on CUDA. And all this is a pain to
install and configure, imagine the shock of your file system and $PATH,
when you install a whole bunch of powerfull instruments, probably never
to use the major part of them ever.

Now it makes sense to do this in a container! Host system doesn't suffer
for no reason, and required install procedures can only be done once,
by the person who maintains the image. Everyone else just pulls it
and starts coding whatever they need. VSCode even provides additional
benifits of "Dev container" which it probably is.

So, good people in ONNX team understand this too and created Dockerfile
to do exactly that! So it is used to provide custom functional for our
needs.

All this is left in deprecatedDkerfile for reference and new Dockerfile
is made with only what we need. Before merging this branch with master,
deprecatedDkerfile should be removed, and the solution must be achieved.

As of QOL improvements: apt-get installs are now cached and ran from
a single RUN command, as to improve on build speed. Some other layers
have also been merged in order to achieve the same goal. Some layers
were left deprecated as they do not provide any functional for now. Also
the order of layers was tweaked in order to be cached according to
good Docker practices.