Stage 1 (installer) added
IMPORTANT!
If image doesn't build, that's intended. Use TWO commands below:
`docker build --target=installer`
`docker build`
To build subsequently
`docker build `

List of changes:
1. Removed repo/test (no longer needed)
2. Added scripts/install-script.sh to manage proper repository cloning
and copying inside of an image
3. Updated Dockerfile

To describe the problem that was solved:
To build ONNX Runtime container official Dockerfile includes
straightforward copying all source code inside build stage. Then
it's built and copied into final stage. But the trial description
demands that repositories are cloned automatically, specifically
from a container (though it does not forbid use of intermediete
container aka layer).

Generally, there are at least 2 approaches to automating cloning
repositories into an image:
1) Script that clones repository, launches build on Dockerfile, and then
builder is instructed to copy Repository inside the image. While this
approach is the best (in a sense that it's faster, takes less memory,
allows much more flexibility and only uses the tools that Docker buildx
fine-tuned to) it contradicts trial's description.
2) Cloning repository during build-time. While it's seemingly the
easiest approach, in reality it is the hardest. It demands a lot of
optimizations, most of which are made impossible by Docker team
themselves. For example, there is an obvious need to avoid cloning
repositories each time building is launched. That demands an access to
host. But Volumes are mounted in runtime, and other mounts exist
only for the lifetime of intermediete container (RUN --mount). Also
docker discards all writes to (--mount=type=bind) out of safety
concerns.

So, currently installer abuses (--mount=type=cache). It clones repoes
into cache-mount and installs them (more in a next paragraph).
Installations are then copied into tester image, but not through
traditional COPY or ADD. It uses RUN since it's the only option to
mount the cache installer made and proceeds to copy installations
through the use of cp programm. All this happens very slow, but still,
it is much faster then cloning everything each build. One more thing
to consider is that copying through RUN cp is implicit from Docker's
perspective so buildx just does not build installer at all, since it
does not see it's connection to tester. It's awesome during consequent
builds, but it requires manually building installer with (--target)
flag in order for repositories to make their way into cache. It must be
done each time Docker decides to destroy cache or if there was none
in the first place.

About install-script.sh:
Basically, it's a wrapper for git-clone. It takes three cmd-line
arguments (directory, name of installation and http to remote-repo).
Then it checks if the installation was already made, if it was
corrupted, and procceds to (re-)install. When installing it just
creates a subdirectory $directory/$name, clones repo in there,
zips it with tar and pipes result to md5sum. The resulting checksum
is then added to $directory/.checksums/$name.md5 for corruption check.
The installation is verified and not reinstalled only if:
1) $directory/$name exists
2) $directory/.checksums/$name.md5 exists
3) Recalculated checksum matches the one written in the file.

The install-script.sh was born out of necessity. When I tried
`RUN --mount=type=cache ... git clone ... ` it yielded git's fatal error
"repository already exists" on each subsequent run except for the first.
Even if this could be solved by just passing some flag to git clone,
Docker handles it's cache in the ways only Docker knows. It could be
destroyed after 48 hours delay or earlier if need be. Or something
could happen to installation and this would make it impossible to build.