Dead end
After finally managing to build (not before repeatedly readjusting a ton
of build.sh flags and deep diving into building process), it turned out
that build is almost equal to an official release. So this branch will
be left as a dead end, because it evidence of my experience and growth
in many related areas.

The new approach is going to be branching from the same base.