v1.0.0 first working build
List of changes:
1. Added samples-commented
2. Added gsc-script
3. Removed custom ENTRYPOINT from Dockerfile since cuda image has it's
    own
4. Complete rework of cmd-script
5. Complete rework of Dockerfile
6. Minor changes to .dockerignore
7. README.md updated

============================= GSC-SCRIPT  ==============================
This script implements Git Sparse Checkout to download samples from
onnxruntime-inference-samples. It downloads only necessery files for the
task, leaving no trace and leftovers.
    Important note 1:
Script downloads content directly into current working directory ($PWD).
In order to use it properly, create new directory, cd into it and
execute script from there (by typing full or relative path to it)
    Important note 2:
If download directory is located inside of a git repository, script
might enforce sparse-checkout onto it. It is recommended to run git
status on said repository after executing gsc-script.sh
    Important note 3:
Script is not used in building Dockerfile. Reasons for this are
described in the next paragraph

========================== SAMPLES-COMMENTED ===========================
For now, it is just the result of executing gsc-script. In future
it will differ from samples in remote repo: according to task
description each line of code should be commented. It wasn't clarified
whether any other changes are allowed, so there will be none.
    Important note 1:
It seems unnecessery to use gsc-script during build, since everything
needed for building the sample should already be present in build
context (except samples-commented/model-explorer) according to task
description.
So the best approach is to leave gsc-script.sh in this project as to
show that it's possible to use it inside the container, while imposing
drastic improvements on build efficiency by replacing it with a simple
COPY.
    Important note 2:
While it may seem more correct to download only squeezenet directory
and build application from there, it is impossible due to flaws in
CMakeLists. It is intented to build all samples using root folder's
CMakeLists which will recursively invoke build to each sample with
correct environment via `add_subdirectory`. This in turn forces us
to include model-explorer in the download, because if we don't,
`add_subdirectory(model-explorer)` in samples-commented/CMakeLists will
raise an error and exit build-process with error code, which in turn
will force Dockerfile build process to exit with error code.

============================= CMD-SCRIPT  ==============================
In it's current state it wraps execution of the application. If the
application exits with error code, then cmd-script.sh prints a list of
issues that were encountered in process of solving the task and offers
possible solutions.

========================== ENTRYPOINT-SCRIPT ===========================
Left as a historic artefact. Cuda container provides it's own entrypoint
which notifies the user whether the container has an access to gpus

============================= DOCKERFILE  ==============================
Build process is now split in two parallel stages:
    1. downloader
    2. builder
    downloader
`downloader` is `scratch`-based. It downloads model and ONNXRuntime
release tarball via Dockerfile ADD command. This happens while builder
updates package lists and installs necessery tools and packages
    builder
Based on cudnn image. It installs gcc and g++ to ensure successful
cmake build, ca-certificates to allow some networking features, and
tensorrt-libs metapackage. In order for squeezenet to run, we're not
required to make full installation of tensorrt, just it's runtime
component (which is tensorrt-libs).
Then unpacks and installs (locally in /usr/local/include and
/usr/local/libs) ONNXRuntime. Afterwards, it removes the leftovers
Then it builds the sample, removes all build artefacts and copies
model inside build directory.
When running resulting image, by default it will invoke cmd-script.sh
and exit.
    Important note 1:
WORKDIR of resulting image is set to /usr/local/src
    Important note 2:
At it's current state, it installs tensorrt 8 (despite the existance of
an official tensorRT image and tensorrt 10). Also it uses cuda 11,
cuDNN 8 and ONNXRuntime 1.14.0 (despite the fact, the existance of newer
versions 12, 9 and 1.19.0 respectively).
Explanation:
    1. When running sample, it looks for lib versions of cuda 11 and
cuDNN 8. This could be solved by creating symlinks, but it is concidered
a bad practice.
    2. Releases of ONNXRuntime are not on the bleeding edge of
technology anyways, just as releases of tensorRT. Their compatibility
is far away from the newest releases of each other and CUDA. Although
this could be mitigated by building ONNXRuntime from source, since it
will adapt to any version of CUDA, cuDNN and tensorRT this way (and
it will probably orchestrate proper cooperation between said tools),
trying this on a outdated machines turns into 30-50 minutes of building
cuDNN files (given that we impose thread restriction on nvcc and jobs
restriction on cmake so that they don't crash the whole process by
running out of RAM).
    3. GPU on the machine this project is developed and tested is no
longer compatible with tensorRT. The last compatible version was 8.5.3.
And it's no longer compatible with newer version of CUDA and cuDNN in
turn.
Given this reasons, using build from source strategy would require
to developing on guess, with testing rendered impossible by hardware
restrictions.